requirements.txt
flask
python-telegram-bot==21.6
pdfplumber
